# AgendaREST/src/main/resources/static/api-docs/swagger.yaml
openapi: '3.0.3'
info:
  title: agenda-java-fullstack
  description: Agenda de contatos.
  version: '1.3.0'
  contact:
    name: Iago Colodetti
    url: https://github.com/iagocolodetti

servers:
  - url: http://localhost:9090/agenda/ws

tags:
  - name: User
    description: REST API para Usuário
  - name: Contact
    description: REST API para Contato

components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header
      
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 1
          description: ID do usuário
          required: false
        username:
          type: string
          example: Nome
          description: Nome de usuário
          required: true
          minimum: 3
          maximum: 45
        password:
          type: string
          example: MinhaSenha123
          description: Senha de acesso
          required: true
          minimum: 5
          maximum: 60

    Contact:
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 1
          description: ID do contato
          required: false
        name:
          type: string
          example: Nome
          description: Nome do contato
          required: true
          minimum: 3
          maximum: 45
        alias:
          type: string
          example: Apelido
          description: Apelido do contato
          required: true
          minimum: 3
          maximum: 45
        phone:
          type: array
          required: true
          minimum: 1
          items:
            $ref: '#/components/schemas/Phone'
        email:
          type: array
          required: true
          minimum: 1
          items:
            $ref: '#/components/schemas/Email'
        deleted:
          type: boolean
          example: false
          description: Contato deletado
          required: false

    Phone:
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 1
          description: ID do telefone
          required: false (for add) / true (for edit)
        phone:
          type: string
          example: '0000-0000'
          description: Telefone do contato
          required: true
          minimum: 7
          maximum: 20
        deleted:
          type: boolean
          example: true
          description: Telefone deletado
          required: false (for add or edit) / true (for delete)
    
    Email:
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 1
          description: ID do e-mail
          required: false (for add) / true (for edit)
        email:
          type: string
          example: 'email@gmail.com'
          description: E-mail do contato
          required: true
          maximum: 60
        deleted:
          type: boolean
          example: true
          description: E-mail deletado
          required: false (for add or edit) / true (for delete)
          
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date
          example: '2020-01-01T12:00:00.000Z'
          description: Data do erro
        status:
          type: integer
          format: int
          example: 500
          description: Código do erro
        error:
          type: string
          example: Internal Server Error
          description: Nome do erro
        message:
          type: string
          example: Não foi possível ...
          description: Mensagem de erro
        path:
          type: string
          example: /exemplo
          description: Caminho usado
        method:
          type: string
          example: POST
          description: Método usado
          
  examples:
    User:
      value:
        username: Nome
        password: MinhaSenha123
      summary: Usuário
      
    NewContact:
      value:
        name: Nome
        alias: Apelido
        phone:
          - phone: '0000-0000'
          - phone: '1111-1111'
        email:
          - email: 'email@gmail.com'
          - email: 'email@hotmail.com'
      summary: Novo Contato

    UpdateContact:
      value:
        name: Nome
        alias: Apelido
        phone:
          - id: 1
            phone: '0000-0000'
          - id: 2
            phone: '1111-1111'
            deleted: true
          - phone: '2222-2222'
        email:
          - id: 1
            email: 'email1@gmail.com'
          - id: 2
            email: 'email2@hotmail.com'
            deleted: true
          - email: 'email3@gmail.com'
      summary: Atualizar Contato
      
    ContactReturn:
      value:
        id: 1
        nome: Nome
        alias: Apelido
        phone:
          - id: 1
            phone: '0000-0000'
          - id: 2
            phone: '1111-1111'
        email:
          - id: 1
            email: email@gmail.com
          - id: 2
            email: email@hotmail.com
      summary: Retorno de Contato
      
    ContactsReturn:
      value:
        - id: 1
          nome: Nome1
          alias: Apelido1
          phone:
            - id: 1
              phone: '0000-0000'
            - id: 2
              phone: '1111-1111'
          email:
            - id: 1
              email: email1@gmail.com
            - id: 2
              email: email2@hotmail.com
        - id: 2
          nome: Nome2
          alias: Apelido2
          phone:
            - id: 3
              phone: '2222-2222'
            - id: 4
              phone: '3333-3333'
          email:
            - id: 3
              email: email3@gmail.com
            - id: 4
              email: email4@hotmail.com
      summary: Retorno de Contatos
  
    Error400:
      value:
        timestamp: '2020-01-01T12:00:00.000Z'
        status: 400
        error: Bad Request
        message: ... não informado
        path: /exemplo
        method: GET
      summary: Erro 400
  
    Error401:
      value:
        timestamp: '2020-01-01T12:00:00.000Z'
        status: 401
        error: Unauthorized
        message: ... não autorizado
        path: /exemplo
        method: GET
      summary: Erro 401
  
    Error404:
      value:
        timestamp: '2020-01-01T12:00:00.000Z'
        status: 404
        error: Not Found
        message: ... não encontrado
        path: /exemplo
        method: GET
      summary: Erro 404
  
    Error409:
      value:
        timestamp: '2020-01-01T12:00:00.000Z'
        status: 409
        error: Conflict
        message: ... já existe
        path: /exemplo
        method: GET
      summary: Erro 409
      
    Error500:
      value:
        timestamp: '2020-01-01T12:00:00.000Z'
        status: 500
        error: Internal Server Erro
        message: Não foi possível ...
        path: /exemplo
        method: GET
      summary: Erro 500
        
paths:
# ------------- User -------------
  /users:
    post:
      tags:
        - User
      summary: Cria um usuário
      description: Cria um usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              User:
                $ref: '#/components/examples/User'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Usuário já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error409'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error500'

# ------------- Session -------------
  /login:
    post:
      tags:
        - User
      summary: Retorna um token de autorização no cabeçalho
      description: Retorna um token de autorização no cabeçalho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              User:
                $ref: '#/components/examples/User'
      responses:
        '200':
          description: Token de autorização retornado no cabeçalho
        '404':
          description: Usuário não encontrado ou senha incorreta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error404'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error500'

# ------------- Contact -------------
  /contacts:
    post:
      tags:
        - Contact
      summary: Adiciona um contato
      description: Adiciona um contato
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Exemplo de Contato
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
            examples:
              Contact:
                $ref: '#/components/examples/NewContact'
      responses:
        '201':
          description: Contato adicionado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
              examples:
                Contact:
                  $ref: '#/components/examples/ContactReturn'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error400'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error401'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error500'

    get:
      tags:
        - Contact
      summary: Retorna uma lista de contatos
      description: Retorna uma lista de contatos
      parameters:
        - in: query
          name: page
          description: Número da página
          required: false
          schema:
            type: integer
        - in: query
          name: size
          description: Tamanho da página (quantidade de itens)
          required: false
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de contatos retornada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
              examples:
                Contacts:
                  $ref: '#/components/examples/ContactsReturn'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error400'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error401'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error500'

  /contacts/{id}:
    put:
      tags:
        - Contact
      summary: Atualiza um contato
      description: Atualiza um contato
      parameters:
        - in: path
          name: id
          description: id do Contato
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: 'Exemplo de Contato (phone / email: com "deleted" para excluir [2] | sem "id" para adicionar [3])'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
            examples:
              Contact:
                $ref: '#/components/examples/UpdateContact'
      responses:
        '204':
          description: Contato atualizado
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error400'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error401'
        '404':
          description: Contato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error404'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Erro:
                  $ref: '#/components/examples/Error500'

    delete:
      tags:
        - Contact
      summary: Exclui um contato
      description: Exclui um contato
      parameters:
        - in: path
          name: id
          description: id do Contato
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Contato excluído
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error400'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error401'
        '404':
          description: Contato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Error:
                  $ref: '#/components/examples/Error404'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Erro:
                  $ref: '#/components/examples/Error500'